requirements.txt
streamlit
yfinance
prophet
pandas
plotly
pillow